move all configparser dependencies to same file
