+++ outdxf.py
@@ -1,30 +1,6 @@
-import configparser
-
 from etik import *
 from KG_dxf import *
-
-
-class DrawingExchangeFormatConfigurations:
-    def __init__(self, path):
-        config = configparser.ConfigParser()
-        config.read(path + "config.kg")
-
-        self.coord_z=config.get('ambiente','esporta coordinate Z dei cerchi')
-        self.cer_livello=str(config.get('stile cerchi','livello'))
-        self.cer_colore=str(config.get('stile cerchi','colore del tratto'))
-        self.ass_colore=str(config.get('stile assi','colore tratto'))
-        self.htesto=str(config.get('stile etichette','altezza caratteri'))
-        self.tes_livello=str(config.get('stile etichette','livello testo'))
-        self.tes_colore=str(config.get('stile etichette','colore testo'))
-        self.angolo=float(config.get('stile etichette','angolazione'))
-        self.etk_livello=config.get('stile etichette','livello')
-        self.etk_colore=str(config.get('stile etichette','colore tratto'))
-        self.etk_dmodA=float(config.get('avanzate','scala dimensioni testo due cifre'))
-        self.etk_dmodB=float(config.get('avanzate','scala dimensioni testo una cifra'))
-        self.etk_dmodAx=float(config.get('avanzate','fattore offset tag due cifre x'))
-        self.etk_dmodAy=float(config.get('avanzate','fattore offset tag due cifre y'))
-        self.etk_dmodBx=float(config.get('avanzate','fattore offset tag una cifra x'))
-        self.etk_dmodBy=float(config.get('avanzate','fattore offset tag una cifra y'))
+from infrastructure import DrawingExchangeFormatConfigurations
 
 
 def esporta_dxf(fori,f, workpath):
